requirements.txt
flask
flask-login
werkzeug
python-pptx
PyMuPDF
textstat
textblob
openai-whisper
torch